% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_Function.R
\name{brew_db_3colors}
\alias{brew_db_3colors}
\title{Create 3 colors}
\usage{
brew_db_3colors(names = NULL)
}
\arguments{
\item{name}{set the names of the three colors}
}
\description{
Create 3 colors
}
