% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewLOI-class.R
\docType{class}
\name{BrewLOI-class}
\alias{BrewLOI-class}
\title{BrewLOI-class}
\description{
Class to store the info in the guess_LOI_table
}
\section{Slots}{

\describe{
\item{\code{snp_count}}{how many SNPs are detected per gene.}

\item{\code{bi_score}}{the average biallelic score of the SNPs detected per gene}

\item{\code{samples}}{the sample names}

\item{\code{genes}}{the gene names}
}}

