% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_ggplot.R
\name{cell_percentage_gradient_blues}
\alias{cell_percentage_gradient_blues}
\title{Blues palette for cell percentages}
\usage{
cell_percentage_gradient_blues(na.value = "grey50")
}
\arguments{
\item{na.value}{color for na}
}
\value{
scale_fill_gradient object with 6 blues
}
\description{
Blues palette for cell percentages
}
