% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_ggplot.R
\name{cell_percentage_gradient_grReds}
\alias{cell_percentage_gradient_grReds}
\title{Green Red palette for cell percentages}
\usage{
cell_percentage_gradient_grReds(na.value = "grey50")
}
\arguments{
\item{na.value}{color for na}
}
\value{
scale_fill_gradient object with 3 green bottom and 3 red up
}
\description{
Green Red palette for cell percentages
}
