% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_Function.R
\name{color_type_gene_label}
\alias{color_type_gene_label}
\title{Associate gene to a color PAR edition
  Function specific of human imprinted gene databases}
\usage{
color_type_gene_label(data_frame_for_dot_plot, gene_annotation)
}
\arguments{
\item{data_frame_for_dot_plot}{a data to plot}

\item{gene_annotation}{annotations}
}
\description{
Associate gene to a color PAR edition
  Function specific of human imprinted gene databases
}
