% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{count_biallelic_samples}
\alias{count_biallelic_samples}
\title{For each gene, counts the biallelic samples}
\usage{
count_biallelic_samples(brew)
}
\arguments{
\item{brew}{a BrewLOI object}
}
\description{
For each gene, counts the biallelic samples
}
