% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_guess_loi_table.R
\name{.remove_weak_snvs}
\alias{.remove_weak_snvs}
\title{Filter tables}
\usage{
.remove_weak_snvs(LOI, thr = 0.2, odepth = 20, min_ac = 3)
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle_v3}

\item{thr}{threshold}

\item{odepth}{overall depth}

\item{min_ac}{minimum allele count}
}
\value{
a GuessLoi object
}
\description{
Filter tables
}
