% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_guess_loi_table.R
\name{extract_chromosomes}
\alias{extract_chromosomes}
\title{Extract sexual chromosomes}
\usage{
extract_chromosomes(LOI, invert_selection = F, chrs = c("X", "Y"))
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle}

\item{invert_selection}{revert selection}

\item{chrs}{vector with chromosomes of interest}
}
\value{
a GuessLoi object
}
\description{
Extract sexual chromosomes
}
