% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_guess_loi_table.R
\name{filter_guess_loi_table}
\alias{filter_guess_loi_table}
\title{Summarize Guess LOI Table object by gene}
\usage{
filter_guess_loi_table(
  LOI,
  keep_type = c(),
  keep_source = c(),
  invert_selection = FALSE
)
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle_v3}

\item{keep_type}{vector with gene type to keep}

\item{invert_selection}{invert the selection}

\item{keep_sources}{vector with gene imprinting sources to keep}
}
\value{
a GuessLoi object
}
\description{
Summarize Guess LOI Table object by gene
}
