% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_gene_average_allelic_ratio}
\alias{plot_gene_average_allelic_ratio}
\title{Plor gene summary table}
\usage{
plot_gene_average_allelic_ratio(summary_table)
}
\arguments{
\item{summary_table}{a object obtained by summarize_table_by_genes}
}
\description{
Plor gene summary table
}
