% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_ASER_related.R
\name{plot_x_false_positive}
\alias{plot_x_false_positive}
\title{Plot the Chromosome coverage and False Positives}
\usage{
plot_x_false_positive(
  asample,
  bin_dim = 1e+06,
  ceil_to = 100,
  hetero_thr = 0.2,
  min_depth = 20,
  min_alt_counts = 3,
  pvalue_thr = NULL,
  chromosomes = c("X"),
  plot_bars = FALSE,
  plot_smoothed_cov = FALSE
)
}
\arguments{
\item{asample}{a samnple derived from gather_info_per_samples}

\item{bin_dim}{binning dimension to compute the coverage}

\item{ceil_to}{upper limit for coverage to increase resolution at low coverage}

\item{hetero_thr}{biallelic threshold}

\item{min_depth}{min depth}

\item{min_alt_counts}{minimal alternative counts}

\item{pvalue_thr}{threshold for pvalue}

\item{chromosomes}{plot chromosomes}

\item{plot_bars}{plot bar coverage}

\item{plot_smoothed_cov}{plot smoothed coverage}
}
\description{
Plot the Chromosome coverage and False Positives
}
