% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_guess_loi_table.R
\name{read_guess_loi_table}
\alias{read_guess_loi_table}
\title{Import Guess LOI Table from file}
\usage{
read_guess_loi_table(file, version = c(4, 3, 2))
}
\arguments{
\item{file}{the path to the file.
File must have 9 annotation columns}
}
\value{
a BrewLoi object
}
\description{
Import Guess LOI Table from file
}
