% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ui_summary_table.R
\name{read_ui_summary_table}
\alias{read_ui_summary_table}
\title{Import Guess LOI Table from file}
\usage{
read_ui_summary_table(file)
}
\arguments{
\item{file}{the path to the summary table file}
}
\value{
a BrewLoi object
}
\description{
Import Guess LOI Table from file
}
