% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_loi_objects.R
\name{rename_samples}
\alias{rename_samples}
\title{Re-name samples of guess loi table}
\usage{
rename_samples(LOI, newNames)
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle_v3}

\item{newNames}{vector with new samples names}
}
\value{
a GuessLoi object
}
\description{
Re-name samples of guess loi table
}
