% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_loi_objects.R
\name{sort_samples}
\alias{sort_samples}
\title{Sort samples of guess loi table}
\usage{
sort_samples(LOI, sorter)
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle_v3}

\item{sorter}{vector with sorted samples names or sorted indexes}
}
\value{
a GuessLoi object
}
\description{
Sort samples of guess loi table
}
