% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_brew_loi_results.R
\name{subset}
\alias{subset}
\title{Sub sample BrewLOI object}
\usage{
subset(brew, sub_samples = NULL, sub_genes = NULL)
}
\arguments{
\item{brew}{a BrewLOI object}

\item{sub_samples}{character vector with the samples to keep}

\item{sub_genes}{character vector with the genes to keep}
}
\value{
a BrewLOI object
}
\description{
Selection can be made on samples or genes or both
}
