% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_table_by_genes.R
\name{summarize_table_by_genes_big}
\alias{summarize_table_by_genes_big}
\title{Summarize Guess LOI Table object by gene}
\usage{
summarize_table_by_genes_big(LOI, thr = 0.2, odepth = 20, min_ac = 3)
}
\arguments{
\item{LOI}{a object obtained by read_guess_loi_tavle_v3}

\item{thr}{threshold}

\item{odepth}{overall depth}

\item{min_ac}{minimum allele count}
}
\value{
a GuessLoi object
}
\description{
Summarize Guess LOI Table object by gene
}
