% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summarizeBrew}
\alias{summarizeBrew}
\title{Annot samples}
\usage{
summarizeBrew(brew, sample_annot)
}
\arguments{
\item{brew}{a BrewLOI object}

\item{sample_annot}{a data frame for sample annotation}
}
\description{
Annot samples
}
